% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AR1.R
\name{AR1}
\alias{AR1}
\title{Autocorrelation Matrix}
\usage{
AR1(rho, n)
}
\arguments{
\item{rho}{The correlation parameter.}

\item{n}{The size of the matrix.}
}
\value{
A n*n matrix with the (i,j) term having value rho^abs(i-j).
}
\description{
This is a function for generating the autocorrelated matrix.
}
\examples{
library(clusterSMART)
AR1(0.5,3)

}
