% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_to_marginal.R
\name{conditional_to_marginal}
\alias{conditional_to_marginal}
\title{Changing Conditional Parameters to Marginal Parameters}
\usage{
conditional_to_marginal(p1, p2, conditional_paras)
}
\arguments{
\item{p1, p2}{The response rate of the first-stage treatment given the first randomization.}

\item{conditional_paras}{A matrix with 6 rows, the first two entries of each row represent the expected mean and variance of each treatment pathway.}
}
\value{
A 4*2 matrix, each row contains the marginal expected mean and variance of each Adaptive Intervention.
}
\description{
This is a function for changing the conditional (treatment-pathway level) parameters to marginal (Adaptive-Intervention level) parameters in a prototypical SMART.
The parameter is defined without considering the baseline covariates.
}
\examples{
library(clusterSMART)
conditional_paras_original=matrix(c(6,8,4,4,4,4,
                                   100,81,64,49,36,36,
                                   0.20,0.18,0.16,0.14,0.12,0.10),6,3)
p1=0.25; p2=0.55
conditional_to_marginal(p1,p2,conditional_paras_original)
}
