% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CS.R
\name{CS}
\alias{CS}
\title{Compound Symmetry (Exchangeable) Matrix}
\usage{
CS(rho, n)
}
\arguments{
\item{rho}{The correlation parameter}

\item{n}{The size of the matrix}
}
\value{
An n*n matrix with diagonal term 1 and off-diagonal term rho
}
\description{
This is a function for generating the compound symmetry (exchangeable) matrix
}
\examples{
library(clusterSMART)
CS(0.5,3)

}
