% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_expand.R
\name{data_expand}
\alias{data_expand}
\title{Expand Data to use in geeglm}
\usage{
data_expand(Y, X = NULL, cluster_id, A1, R, A2, seperate = FALSE)
}
\arguments{
\item{Y}{Length n numeric vector or n*1 matrix of the individual outcome.}

\item{X}{NULL, length n vector or n*p matrix for the baseline.}

\item{cluster_id}{n*1 numeric vector of the cluster id which the individual belongs to.}

\item{A1}{Length n vector with value in {-1,1}, represents the result of the cluster-level randomization for the first-stage treatment.}

\item{R}{Length n vector with value in {0,1}, represents the cluster-level response status to the first-stage treatment.}

\item{A2}{Length n vector with values in {-1,1,NA}, represents the result of the cluster-level randomization for the second-stage treatment.}

\item{seperate}{Logical. If TRUE then each response cluster is seperated into two clusters, if FALSE the response cluster is expanded to be twice large. Default is FALSE.}
}
\value{
A list of following two objects:\tabular{ll}{
 \code{data}\tab The expanded dataset.\cr
 \tab \cr
 \code{structure}\tab If cluster sizes are not equal or seperate is TRUE, NULL; if cluster sizes are equal, a userdefined correlation structure that can be used in the 'zcor' entry in geeglm. \cr
 }
}
\description{
This is a function for doing data expansion trick to clustered SMART data to make it available for using off-the-shelf geeglm function to do proper inference.\cr
Such analysis is only valid when using independence working variance structure, or when the cluster sizes are equal and use the customed variance structure for exchangeable.
}
\examples{
library(clusterSMART)
library(geepack)
conditional_paras_original=matrix(c(6,8,4,4,4,4,
                                   100,81,64,49,36,36,
                                   0.20,0.18,0.16,0.14,0.12,0.10),6,3)
p1=0.25; p2=0.55
effectsize_dtr=0.5; rhoadjust=1
effectsize_eta=NULL; sigma2_x=NULL
paras=parameter_adjust(conditional_paras_original=conditional_paras_original,
                      p1=p1,p2=p2,aimed_comparison=c(1,0,0,1),
                      effectsize_dtr=effectsize_dtr,effectsize_eta=effectsize_eta,
                      rhoadjust=rhoadjust,sigma2_x=sigma2_x)
temp=generate_SMART(conditional_paras=paras$conditional_paras,
                  p1=0.25,p2=0.55,N=50,Mmax=3,Mmin=3,seed=1234)

expanded=data_expand(Y=temp$data$Y,cluster_id=temp$data$id,
                    A1=temp$data$A1,R=temp$data$R,A2=temp$data$A2,seperate=FALSE)
model1=geeglm(Y~A1+A2+I(A1*A2),id=id,family=gaussian(),weight=W,data=expanded$data,
             corstr="userdefined",zcor=expanded$structure)
summary(model1)

}
