% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_testing.R
\name{hypothesis_testing}
\alias{hypothesis_testing}
\title{General Hypothesis Testing}
\usage{
hypothesis_testing(result, aimed_test, alpha = 0.05, use_t = TRUE)
}
\arguments{
\item{result}{List output of solve_SMART function, containing details of parameter estimation and variance estiamtor.}

\item{aimed_test}{r*(q+p+1) matrix, each row corresponds to a linear combination of coefficients that is checked in hypothesis. q=4 is number of Adaptive Interventions, p is number of baseline covariates, the last row is constant term.}

\item{alpha}{The significance level (in inference we construct the (1-alpha) confidence interval).}

\item{use_t}{Logical. If TRUE uses t-distribution instead of normal when constructing the confidence interval and calculating the p-value.}
}
\value{
A matrix of test result.'
}
\description{
This is a function for doing general hypothesis testing with given solve_SMART output.
}
\examples{
library(clusterSMART)
library(geepack)
conditional_paras_original=matrix(c(6,8,4,4,4,4,
                                   100,81,64,49,36,36,
                                   0.20,0.18,0.16,0.14,0.12,0.10),6,3)
p1=0.25; p2=0.55
effectsize_dtr=0.5; rhoadjust=1
effectsize_eta=c(0.2,0.5); sigma2_x=c(1,2)
paras=parameter_adjust(conditional_paras_original=conditional_paras_original,
                      p1=p1,p2=p2,aimed_comparison=c(1,0,0,1),
                      effectsize_dtr=effectsize_dtr,effectsize_eta=effectsize_eta,
                      rhoadjust=rhoadjust,sigma2_x=sigma2_x)
temp=generate_SMART(conditional_paras=paras$conditional_paras,
                   eta_x=paras$eta,sigma2_x=c(1,2),detail_x=c(1,1),
                   p1=0.25,p2=0.55,N=500,Mmax=5,Mmin=2,seed=1234)
res=solve_SMART(Y=temp$data$Y,X=temp$data$X,cluster_id=temp$data$id,
               A1=temp$data$A1,R=temp$data$R,A2=temp$data$A2,
               aimed_comparison=matrix(c(1,0,0,1,1,0,1,0),2,4,byrow=TRUE),
               variance_structure=0,correlation_structure=1,ICC_lower_thresh=0,
               max_iter=10,convergence_thresh=1e-5,alpha=0.05,
               estimate_weight=FALSE,dof_adjustment=FALSE,use_t=FALSE,
               bias_correction=FALSE,verbose=2)
hypothesis_testing(res,aimed_test=matrix(c(1,-1,0,0,1,0,-5,
                                          1,0,0,0,0,-1,1),2,7,byrow=TRUE),
                  alpha=0.05,use_t=TRUE)

}
