% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_to_response.R
\name{pathway_to_response}
\alias{pathway_to_response}
\title{Change from Treatment Pathway to (A1,R,A2)}
\usage{
pathway_to_response(L)
}
\arguments{
\item{L}{Length n vector with values in {1,2,...,6}, represents the label of treatment pathway.}
}
\value{
A n*3 matrix of the extracted results:\tabular{ll}{
 \code{A1}\tab The result of the cluster-level randomization for the first-stage treatment.\cr
 \tab \cr
 \code{R}\tab The cluster-level response status to the first-stage treatment. \cr
 \tab \cr
 \code{A2}\tab The result of the cluster-level randomization for the second-stage treatment.\cr
 }
}
\description{
This is a function for extracting the randomization result and response status (A1,R,A2) from treatment pathway label under prototypical SMART.
}
\examples{
library(clusterSMART)
L=c(1:6)
pathway_to_response(L)

}
