% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_SMART.R
\name{solve_SMART}
\alias{solve_SMART}
\title{Solve clustered SMART}
\usage{
solve_SMART(
  Y,
  X = NULL,
  cluster_id,
  A1,
  R,
  A2,
  aimed_comparison = matrix(c(1, 1, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0, 0,
    1, 0, 1, 1), 6, 4),
  variance_structure = 1,
  correlation_structure = 2,
  ICC_lower_thresh = 0,
  max_iter = 10,
  convergence_thresh = 1e-05,
  alpha = 0.05,
  estimate_weight = FALSE,
  dof_adjustment = TRUE,
  use_t = TRUE,
  bias_correction = TRUE,
  verbose = 2
)
}
\arguments{
\item{Y}{Length n numeric vector or n*1 matrix of the individual outcome.}

\item{X}{NULL, length n vector or n*p matrix for the baseline.}

\item{cluster_id}{n*1 numeric vector of the cluster id which the individual belongs to.}

\item{A1}{Length n vector with value in {-1,1}, represents the result of the cluster-level randomization for the first-stage treatment.}

\item{R}{Length n vector with value in {0,1}, represents the cluster-level response status to the first-stage treatment.}

\item{A2}{Length n vector with values in {-1,1,NA}, represents the result of the cluster-level randomization for the second-stage treatment.}

\item{aimed_comparison}{NULL, r*4 matrix, each row has 2 0's and 2 1's, corresponding to the proposed comparison between AIs (default is all pairs of comparison).}

\item{variance_structure}{The marginal variance in the working model, 0 for same variance across Adaptive Intervention, 1 for different variance (default = 1).}

\item{correlation_structure}{The marginal intra-cluster correlation in the working model, 0 for independence, 1 for same ICC across AI, 2 for different ICC (default = 2).}

\item{ICC_lower_thresh}{A value in [-1,1] representing the set lower cutting threshold for ICC in working model, default is 0, setting to -1 means no cutting.}

\item{max_iter}{A natural integer representing the max allowed iteration in solving the effect estimates.}

\item{convergence_thresh}{A positive value representing the convergence threshold in the iterative algorithm of solving the effect estiamtes.}

\item{alpha}{The significance level (in inference we construct the (1-alpha) confidence interval).}

\item{estimate_weight}{Logical. If TRUE uses empirical estiamted inverse probability weight and adjust for the uncertainty introduced in the estimation in the variance estimator.}

\item{dof_adjustment}{Logical. If TRUE uses degree-of-freedom adjustment in the variance estiamtor.}

\item{use_t}{Logical. If TRUE uses t-distribution instead of normal when constructing the confidence interval and calculating the p-value.}

\item{bias_correction}{Logical. If TRUE uses the bias-corrected variance estimator for small-sample inference.}

\item{verbose}{Numeric value in {0,1,2,3} representing the printed output format. 0 for printing no output, 1 for coefficient estimation, 2 for coefficient estimation and working variance parameters, 3 for most detailed output including aimed comparison and variance estiamtor.}
}
\value{
A list of result of analysis:\tabular{ll}{
 \code{summary_paras}\tab A list of the result of coefficient estimate.\cr
 \tab \cr
 \code{marginal_paras}\tab A 4*3 numeric matrix of the marginal mean,variance and ICC of each Adaptive Intervention.\cr
 \tab \cr
 \code{var_estimator}\tab The variance estimator of the coefficients.\cr
 \tab \cr
 \code{formula}\tab The formula of marginal mean regression model.\cr
 \tab \cr
 \code{comparison}\tab NULL or a list of the result of the aimed comparison between Adaptive Interventions.\cr
 \tab \cr
 \code{N}\tab Number of clusters.\cr
 }
}
\description{
This is a function for solving clustered SMART data and do inference.\cr
This function uses iterative algorithm to get the coefficient estimates for main effects and covaraites, and uses sandwich estimator to get proper inference.\cr
Also, this function is flexible to use different types of working variance model and implement different small-sample modification of variance estiamtors.
}
\examples{
library(clusterSMART)
conditional_paras_original=matrix(c(6,8,4,4,4,4,100,81,64,49,36,36,0.20,0.18,0.16,0.14,0.12,0.10),6,3)
p1=0.25; p2=0.55
effectsize_dtr=0.5; rhoadjust=1
effectsize_eta=c(0.2,0.5); sigma2_x=c(1,2)
paras=parameter_adjust(conditional_paras_original=conditional_paras_original,p1=p1,p2=p2,aimed_comparison=c(1,0,0,1),effectsize_dtr=effectsize_dtr,effectsize_eta=effectsize_eta,rhoadjust[rhoadjust],sigma2_x=sigma2_x)
temp=generate_SMART(conditional_paras=paras$conditional_paras,eta_x=paras$eta,sigma2_x=c(1,2),detail_x=c(1,1),p1=0.25,p2=0.55,N=500,Mmax=10,Mmin=1,seed=1234)
res=solve_SMART(Y=temp$data$Y,X=temp$data$X,cluster_id=temp$data$id,A1=temp$data$A1,R=temp$data$R,A2=temp$data$A2,
                aimed_comparison=matrix(c(1,0,0,1,1,0,1,0),2,4,byrow=TRUE),
                variance_structure=0,correlation_structure=1,ICC_lower_thresh=0,
                max_iter=10,convergence_thresh=1e-5,alpha=0.05,
                estimate_weight=FALSE,dof_adjustment=FALSE,use_t=FALSE,bias_correction=FALSE,verbose=2)

}
